requirements.txt
torch
numpy
tqdm
torchvision
pycocotools
pillow
CLIP from here: https://github.com/openai/CLIP